+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Upcoming
+### Added
+- Accept semver keywords as versions
+
 ## v2.3.0 - 2015-03-11
 ### Added
 - Validate the version string